@@ -3,9 +3,10 @@ description: Flutter plugin enabling simple passkey authentication. Can be eithe
 homepage: https://docs.corbado.com/overview/welcome
 repository: https://github.com/corbado/flutter-passkeys/tree/main/packages/passkeys/passkeys
 version: 2.17.4
+publish_to: none
 
 environment:
-  sdk: ">=3.0.0 <4.0.0"
+  sdk: '>=3.8.0 <4.0.0'
   flutter: ">=3.0.0"
 
 flutter:
@@ -27,12 +28,18 @@ dependencies:
   flutter:
     sdk: flutter
   json_annotation: ^4.8.1
-  passkeys_android: ^2.11.0
-  passkeys_darwin: ^0.3.0
-  passkeys_doctor: ^1.2.0
-  passkeys_platform_interface: ^2.6.0
-  passkeys_web: ^2.8.1
-  passkeys_windows: ^0.1.1
+  passkeys_android: 
+    path: ../passkeys_android
+  passkeys_darwin: 
+    path: ../passkeys_darwin
+  passkeys_doctor: 
+    path: ../passkeys_doctor
+  passkeys_platform_interface:
+    path: ../passkeys_platform_interface
+  passkeys_web: 
+    path: ../passkeys_web
+  passkeys_windows: 
+    path: ../passkeys_windows
   ua_client_hints: ^1.1.3
 
 dev_dependencies:
@@ -43,3 +50,7 @@ dev_dependencies:
   mocktail: ^1.0.0
   plugin_platform_interface: ^2.0.0
   very_good_analysis: ^5.0.0
+  
+dependency_overrides:
+  passkeys_platform_interface:
+    path: ../passkeys_platform_interface
