@@ -219,6 +219,7 @@ class RegisterResponse {
     required this.clientDataJSON,
     required this.attestationObject,
     required this.transports,
+    this.clientExtensionResults,
   });
 
   /// The ID
@@ -236,6 +237,9 @@ class RegisterResponse {
   /// The supported transports for the authenticator
   List<String?> transports;
 
+  /// The clientExtensionResults - PRF results
+  Map<String?, Object?>? clientExtensionResults;
+
   Object encode() {
     return <Object?>[
       id,
@@ -243,6 +247,7 @@ class RegisterResponse {
       clientDataJSON,
       attestationObject,
       transports,
+      clientExtensionResults,
     ];
   }
 
@@ -254,6 +259,7 @@ class RegisterResponse {
       clientDataJSON: result[2]! as String,
       attestationObject: result[3]! as String,
       transports: (result[4] as List<Object?>?)!.cast<String?>(),
+      clientExtensionResults: (result[5] as Map<Object?, Object?>?)?.cast<String?, Object?>(),
     );
   }
 }
@@ -267,6 +273,7 @@ class AuthenticateResponse {
     required this.authenticatorData,
     required this.signature,
     required this.userHandle,
+    this.clientExtensionResults,
   });
 
   /// The ID
@@ -284,7 +291,11 @@ class AuthenticateResponse {
   /// The signature
   String signature;
 
+  /// The userHandle
   String userHandle;
+
+  /// The clientExtensionResults - PRF results
+  Map<String?, Object?>? clientExtensionResults;
 
   Object encode() {
     return <Object?>[
@@ -294,6 +305,7 @@ class AuthenticateResponse {
       authenticatorData,
       signature,
       userHandle,
+      clientExtensionResults,
     ];
   }
 
@@ -306,6 +318,7 @@ class AuthenticateResponse {
       authenticatorData: result[3]! as String,
       signature: result[4]! as String,
       userHandle: result[5]! as String,
+      clientExtensionResults: (result[6] as Map<Object?, Object?>?)?.cast<String?, Object?>(),
     );
   }
 }
@@ -346,21 +359,21 @@ class _PasskeysApiCodec extends StandardMessageCodec {
   @override
   Object? readValueOfType(int type, ReadBuffer buffer) {
     switch (type) {
-      case 128: 
+      case 128:
         return AllowCredential.decode(readValue(buffer)!);
-      case 129: 
+      case 129:
         return AuthenticateResponse.decode(readValue(buffer)!);
-      case 130: 
+      case 130:
         return AuthenticatorSelection.decode(readValue(buffer)!);
-      case 131: 
+      case 131:
         return ExcludeCredential.decode(readValue(buffer)!);
-      case 132: 
+      case 132:
         return PubKeyCredParam.decode(readValue(buffer)!);
-      case 133: 
+      case 133:
         return RegisterResponse.decode(readValue(buffer)!);
-      case 134: 
+      case 134:
         return RelyingParty.decode(readValue(buffer)!);
-      case 135: 
+      case 135:
         return User.decode(readValue(buffer)!);
       default:
         return super.readValueOfType(type, buffer);
@@ -372,18 +385,15 @@ class PasskeysApi {
   /// Constructor for [PasskeysApi].  The [binaryMessenger] named argument is
   /// available for dependency injection.  If it is left null, the default
   /// BinaryMessenger will be used which routes to the host platform.
-  PasskeysApi({BinaryMessenger? binaryMessenger})
-      : _binaryMessenger = binaryMessenger;
+  PasskeysApi({BinaryMessenger? binaryMessenger}) : _binaryMessenger = binaryMessenger;
   final BinaryMessenger? _binaryMessenger;
 
   static const MessageCodec<Object?> codec = _PasskeysApiCodec();
 
   Future<bool> canAuthenticate() async {
-    final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
-        'dev.flutter.pigeon.passkeys_android.PasskeysApi.canAuthenticate', codec,
-        binaryMessenger: _binaryMessenger);
-    final List<Object?>? replyList =
-        await channel.send(null) as List<Object?>?;
+    final BasicMessageChannel<Object?> channel =
+        BasicMessageChannel<Object?>('dev.flutter.pigeon.passkeys_android.PasskeysApi.canAuthenticate', codec, binaryMessenger: _binaryMessenger);
+    final List<Object?>? replyList = await channel.send(null) as List<Object?>?;
     if (replyList == null) {
       throw PlatformException(
         code: 'channel-error',
@@ -406,11 +416,9 @@ class PasskeysApi {
   }
 
   Future<bool> hasPasskeySupport() async {
-    final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
-        'dev.flutter.pigeon.passkeys_android.PasskeysApi.hasPasskeySupport', codec,
-        binaryMessenger: _binaryMessenger);
-    final List<Object?>? replyList =
-        await channel.send(null) as List<Object?>?;
+    final BasicMessageChannel<Object?> channel =
+        BasicMessageChannel<Object?>('dev.flutter.pigeon.passkeys_android.PasskeysApi.hasPasskeySupport', codec, binaryMessenger: _binaryMessenger);
+    final List<Object?>? replyList = await channel.send(null) as List<Object?>?;
     if (replyList == null) {
       throw PlatformException(
         code: 'channel-error',
@@ -432,12 +440,29 @@ class PasskeysApi {
     }
   }
 
-  Future<RegisterResponse> register(String arg_challenge, RelyingParty arg_relyingParty, User arg_user, AuthenticatorSelection? arg_authenticatorSelection, List<PubKeyCredParam?>? arg_pubKeyCredParams, int? arg_timeout, String? arg_attestation, List<ExcludeCredential?> arg_excludeCredentials) async {
-    final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
-        'dev.flutter.pigeon.passkeys_android.PasskeysApi.register', codec,
-        binaryMessenger: _binaryMessenger);
-    final List<Object?>? replyList =
-        await channel.send(<Object?>[arg_challenge, arg_relyingParty, arg_user, arg_authenticatorSelection, arg_pubKeyCredParams, arg_timeout, arg_attestation, arg_excludeCredentials]) as List<Object?>?;
+  Future<RegisterResponse> register(
+      String arg_challenge,
+      RelyingParty arg_relyingParty,
+      User arg_user,
+      AuthenticatorSelection? arg_authenticatorSelection,
+      List<PubKeyCredParam?>? arg_pubKeyCredParams,
+      int? arg_timeout,
+      String? arg_attestation,
+      List<ExcludeCredential?> arg_excludeCredentials,
+      String? arg_salt) async {
+    final BasicMessageChannel<Object?> channel =
+        BasicMessageChannel<Object?>('dev.flutter.pigeon.passkeys_android.PasskeysApi.register', codec, binaryMessenger: _binaryMessenger);
+    final List<Object?>? replyList = await channel.send(<Object?>[
+      arg_challenge,
+      arg_relyingParty,
+      arg_user,
+      arg_authenticatorSelection,
+      arg_pubKeyCredParams,
+      arg_timeout,
+      arg_attestation,
+      arg_excludeCredentials,
+      arg_salt
+    ]) as List<Object?>?;
     if (replyList == null) {
       throw PlatformException(
         code: 'channel-error',
@@ -459,12 +484,20 @@ class PasskeysApi {
     }
   }
 
-  Future<AuthenticateResponse> authenticate(String arg_relyingPartyId, String arg_challenge, int? arg_timeout, String? arg_userVerification, List<AllowCredential?>? arg_allowCredentials, bool? arg_preferImmediatelyAvailableCredentials) async {
-    final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
-        'dev.flutter.pigeon.passkeys_android.PasskeysApi.authenticate', codec,
-        binaryMessenger: _binaryMessenger);
-    final List<Object?>? replyList =
-        await channel.send(<Object?>[arg_relyingPartyId, arg_challenge, arg_timeout, arg_userVerification, arg_allowCredentials, arg_preferImmediatelyAvailableCredentials]) as List<Object?>?;
+  Future<AuthenticateResponse> authenticate(String arg_relyingPartyId, String arg_challenge, int? arg_timeout, String? arg_userVerification,
+      List<AllowCredential?>? arg_allowCredentials, bool? arg_preferImmediatelyAvailableCredentials,
+      {String? arg_salt}) async {
+    final BasicMessageChannel<Object?> channel =
+        BasicMessageChannel<Object?>('dev.flutter.pigeon.passkeys_android.PasskeysApi.authenticate', codec, binaryMessenger: _binaryMessenger);
+    final List<Object?>? replyList = await channel.send(<Object?>[
+      arg_relyingPartyId,
+      arg_challenge,
+      arg_timeout,
+      arg_userVerification,
+      arg_allowCredentials,
+      arg_preferImmediatelyAvailableCredentials,
+      arg_salt
+    ]) as List<Object?>?;
     if (replyList == null) {
       throw PlatformException(
         code: 'channel-error',
@@ -490,8 +523,7 @@ class PasskeysApi {
     final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
         'dev.flutter.pigeon.passkeys_android.PasskeysApi.cancelCurrentAuthenticatorOperation', codec,
         binaryMessenger: _binaryMessenger);
-    final List<Object?>? replyList =
-        await channel.send(null) as List<Object?>?;
+    final List<Object?>? replyList = await channel.send(null) as List<Object?>?;
     if (replyList == null) {
       throw PlatformException(
         code: 'channel-error',
