@@ -109,6 +109,7 @@ class RegisterResponse {
     required this.clientDataJSON,
     required this.attestationObject,
     required this.transports,
+    this.clientExtensionResults,
   });
 
   /// The ID
@@ -126,6 +127,9 @@ class RegisterResponse {
   /// The supported transports for the authenticator
   List<String?> transports;
 
+  /// The clientExtensionResults - PRF results
+  Map<String?, Object?>? clientExtensionResults;
+
   Object encode() {
     return <Object?>[
       id,
@@ -133,6 +137,7 @@ class RegisterResponse {
       clientDataJSON,
       attestationObject,
       transports,
+      clientExtensionResults,
     ];
   }
 
@@ -144,6 +149,7 @@ class RegisterResponse {
       clientDataJSON: result[2]! as String,
       attestationObject: result[3]! as String,
       transports: (result[4] as List<Object?>?)!.cast<String?>(),
+      clientExtensionResults: (result[5] as Map<Object?, Object?>?)?.cast<String?, Object?>(),
     );
   }
 }
@@ -157,6 +163,7 @@ class AuthenticateResponse {
     required this.authenticatorData,
     required this.signature,
     this.userHandle,
+    this.clientExtensionResults,
   });
 
   /// The ID
@@ -176,6 +183,9 @@ class AuthenticateResponse {
 
   String? userHandle;
 
+  /// The clientExtensionResults - PRF results
+  Map<String?, Object?>? clientExtensionResults;
+
   Object encode() {
     return <Object?>[
       id,
@@ -184,6 +194,7 @@ class AuthenticateResponse {
       authenticatorData,
       signature,
       userHandle,
+      clientExtensionResults,
     ];
   }
 
@@ -196,6 +207,7 @@ class AuthenticateResponse {
       authenticatorData: result[3]! as String,
       signature: result[4]! as String,
       userHandle: result[5] as String?,
+      clientExtensionResults: (result[6] as Map<Object?, Object?>?)?.cast<String?, Object?>(),
     );
   }
 }
@@ -227,15 +239,15 @@ class _PasskeysApiCodec extends StandardMessageCodec {
   @override
   Object? readValueOfType(int type, ReadBuffer buffer) {
     switch (type) {
-      case 128: 
+      case 128:
         return AuthenticateResponse.decode(readValue(buffer)!);
-      case 129: 
+      case 129:
         return CredentialType.decode(readValue(buffer)!);
-      case 130: 
+      case 130:
         return RegisterResponse.decode(readValue(buffer)!);
-      case 131: 
+      case 131:
         return RelyingParty.decode(readValue(buffer)!);
-      case 132: 
+      case 132:
         return User.decode(readValue(buffer)!);
       default:
         return super.readValueOfType(type, buffer);
@@ -247,18 +259,15 @@ class PasskeysApi {
   /// Constructor for [PasskeysApi].  The [binaryMessenger] named argument is
   /// available for dependency injection.  If it is left null, the default
   /// BinaryMessenger will be used which routes to the host platform.
-  PasskeysApi({BinaryMessenger? binaryMessenger})
-      : _binaryMessenger = binaryMessenger;
+  PasskeysApi({BinaryMessenger? binaryMessenger}) : _binaryMessenger = binaryMessenger;
   final BinaryMessenger? _binaryMessenger;
 
   static const MessageCodec<Object?> codec = _PasskeysApiCodec();
 
   Future<bool> canAuthenticate() async {
-    final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
-        'dev.flutter.pigeon.passkeys_darwin.PasskeysApi.canAuthenticate', codec,
-        binaryMessenger: _binaryMessenger);
-    final List<Object?>? replyList =
-        await channel.send(null) as List<Object?>?;
+    final BasicMessageChannel<Object?> channel =
+        BasicMessageChannel<Object?>('dev.flutter.pigeon.passkeys_darwin.PasskeysApi.canAuthenticate', codec, binaryMessenger: _binaryMessenger);
+    final List<Object?>? replyList = await channel.send(null) as List<Object?>?;
     if (replyList == null) {
       throw PlatformException(
         code: 'channel-error',
@@ -281,11 +290,9 @@ class PasskeysApi {
   }
 
   Future<bool> hasBiometrics() async {
-    final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
-        'dev.flutter.pigeon.passkeys_darwin.PasskeysApi.hasBiometrics', codec,
-        binaryMessenger: _binaryMessenger);
-    final List<Object?>? replyList =
-        await channel.send(null) as List<Object?>?;
+    final BasicMessageChannel<Object?> channel =
+        BasicMessageChannel<Object?>('dev.flutter.pigeon.passkeys_darwin.PasskeysApi.hasBiometrics', codec, binaryMessenger: _binaryMessenger);
+    final List<Object?>? replyList = await channel.send(null) as List<Object?>?;
     if (replyList == null) {
       throw PlatformException(
         code: 'channel-error',
@@ -307,12 +314,31 @@ class PasskeysApi {
     }
   }
 
-  Future<RegisterResponse> register(String arg_challenge, RelyingParty arg_relyingParty, User arg_user, List<CredentialType?> arg_excludeCredentials, List<int?> arg_pubKeyCredValues, bool arg_canBePlatformAuthenticator, bool arg_canBeSecurityKey, String? arg_residentKeyPreference, String? arg_attestationPreference) async {
-    final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
-        'dev.flutter.pigeon.passkeys_darwin.PasskeysApi.register', codec,
-        binaryMessenger: _binaryMessenger);
-    final List<Object?>? replyList =
-        await channel.send(<Object?>[arg_challenge, arg_relyingParty, arg_user, arg_excludeCredentials, arg_pubKeyCredValues, arg_canBePlatformAuthenticator, arg_canBeSecurityKey, arg_residentKeyPreference, arg_attestationPreference]) as List<Object?>?;
+  Future<RegisterResponse> register(
+      String arg_challenge,
+      RelyingParty arg_relyingParty,
+      User arg_user,
+      List<CredentialType?> arg_excludeCredentials,
+      List<int?> arg_pubKeyCredValues,
+      bool arg_canBePlatformAuthenticator,
+      bool arg_canBeSecurityKey,
+      String? arg_residentKeyPreference,
+      String? arg_attestationPreference,
+      String? arg_salt) async {
+    final BasicMessageChannel<Object?> channel =
+        BasicMessageChannel<Object?>('dev.flutter.pigeon.passkeys_darwin.PasskeysApi.register', codec, binaryMessenger: _binaryMessenger);
+    final List<Object?>? replyList = await channel.send(<Object?>[
+      arg_challenge,
+      arg_relyingParty,
+      arg_user,
+      arg_excludeCredentials,
+      arg_pubKeyCredValues,
+      arg_canBePlatformAuthenticator,
+      arg_canBeSecurityKey,
+      arg_residentKeyPreference,
+      arg_attestationPreference,
+      arg_salt
+    ]) as List<Object?>?;
     if (replyList == null) {
       throw PlatformException(
         code: 'channel-error',
@@ -334,12 +360,13 @@ class PasskeysApi {
     }
   }
 
-  Future<AuthenticateResponse> authenticate(String arg_relyingPartyId, String arg_challenge, bool arg_conditionalUI, List<CredentialType?> arg_allowedCredentials, bool arg_preferImmediatelyAvailableCredentials) async {
-    final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
-        'dev.flutter.pigeon.passkeys_darwin.PasskeysApi.authenticate', codec,
-        binaryMessenger: _binaryMessenger);
-    final List<Object?>? replyList =
-        await channel.send(<Object?>[arg_relyingPartyId, arg_challenge, arg_conditionalUI, arg_allowedCredentials, arg_preferImmediatelyAvailableCredentials]) as List<Object?>?;
+  Future<AuthenticateResponse> authenticate(String arg_relyingPartyId, String arg_challenge, bool arg_conditionalUI,
+      List<CredentialType?> arg_allowedCredentials, bool arg_preferImmediatelyAvailableCredentials, String? arg_salt) async {
+    final BasicMessageChannel<Object?> channel =
+        BasicMessageChannel<Object?>('dev.flutter.pigeon.passkeys_darwin.PasskeysApi.authenticate', codec, binaryMessenger: _binaryMessenger);
+    final List<Object?>? replyList = await channel
+            .send(<Object?>[arg_relyingPartyId, arg_challenge, arg_conditionalUI, arg_allowedCredentials, arg_preferImmediatelyAvailableCredentials, arg_salt])
+        as List<Object?>?;
     if (replyList == null) {
       throw PlatformException(
         code: 'channel-error',
@@ -365,8 +392,7 @@ class PasskeysApi {
     final BasicMessageChannel<Object?> channel = BasicMessageChannel<Object?>(
         'dev.flutter.pigeon.passkeys_darwin.PasskeysApi.cancelCurrentAuthenticatorOperation', codec,
         binaryMessenger: _binaryMessenger);
-    final List<Object?>? replyList =
-        await channel.send(null) as List<Object?>?;
+    final List<Object?>? replyList = await channel.send(null) as List<Object?>?;
     if (replyList == null) {
       throw PlatformException(
         code: 'channel-error',
