@@ -58,6 +58,7 @@ class RegisterResponse {
     required this.clientDataJSON,
     required this.attestationObject,
     required this.transports,
+    this.clientExtensionResults = const {},
   });
 
   /// The ID
@@ -74,6 +75,9 @@ class RegisterResponse {
 
   /// The supported transports for the authenticator
   final List<String?> transports;
+
+  /// The clientExtensionResults - PRF results
+  final Map<String?, Object?>? clientExtensionResults;
 }
 
 /// Represents an authenticate response
@@ -86,6 +90,7 @@ class AuthenticateResponse {
     required this.authenticatorData,
     required this.signature,
     this.userHandle,
+    this.clientExtensionResults = const {},
   });
 
   /// The ID
@@ -104,6 +109,9 @@ class AuthenticateResponse {
   final String signature;
 
   final String? userHandle;
+
+  /// The clientExtensionResults - PRF results
+  final Map<String?, Object?>? clientExtensionResults;
 }
 
 @HostApi()
@@ -123,6 +131,7 @@ abstract class PasskeysApi {
     bool canBeSecurityKey,
     String? residentKeyPreference,
     String? attestationPreference,
+    String? salt,
   );
 
   @async
@@ -131,8 +140,9 @@ abstract class PasskeysApi {
     String challenge,
     bool conditionalUI,
     List<CredentialType> allowedCredentials,
-    bool preferImmediatelyAvailableCredentials,
-  );
+    bool preferImmediatelyAvailableCredentials, {
+    String? salt,
+  });
 
   @async
   void cancelCurrentAuthenticatorOperation();
