@@ -40,14 +40,12 @@ abstract class PasskeysPlatform extends PlatformInterface {
   /// Handles the platform-specific steps for the registration flow
   /// (see https://webauthn.guide/#registration)
   /// Namely it creates a public/private key pair (only the public key will be returned)
-  Future<RegisterResponseType> register(RegisterRequestType request);
+  Future<RegisterResponseType> register(RegisterRequestType request, String salt);
 
   /// Handles the platform-specific steps for the authentication flow
   /// (see https://webauthn.guide/#authentication)
   /// Namely it creates a signature for the challenge issued by the relying party
-  Future<AuthenticateResponseType> authenticate(
-    AuthenticateRequestType request,
-  );
+  Future<AuthenticateResponseType> authenticate(AuthenticateRequestType request, {String? salt});
 
   /// Cancels the ongoing authenticator operation (if there is one).
   /// This is important for the case when conditional UI has been started but
