@@ -2,13 +2,15 @@ import 'dart:convert';
 
 class AuthenticateResponseType {
   /// Constructs a new instance.
-  const AuthenticateResponseType(
-      {required this.id,
-      required this.rawId,
-      required this.clientDataJSON,
-      required this.authenticatorData,
-      required this.signature,
-      required this.userHandle});
+  const AuthenticateResponseType({
+    required this.id,
+    required this.rawId,
+    required this.clientDataJSON,
+    required this.authenticatorData,
+    required this.signature,
+    required this.userHandle,
+    this.clientExtensionResults,
+  });
 
   /// Constructs a new instance from a JSON string.
   factory AuthenticateResponseType.fromJsonString(String jsonString) {
@@ -27,8 +29,7 @@ class AuthenticateResponseType {
     if (json.containsKey('response')) {
       final response = json['response'];
       if (response is! Map<String, dynamic>) {
-        throw FormatException(
-            'Expected "response" to be a Map, got ${response.runtimeType}');
+        throw FormatException('Expected "response" to be a Map, got ${response.runtimeType}');
       }
       return AuthenticateResponseType(
         id: json['id'] as String? ?? '',
@@ -37,6 +38,7 @@ class AuthenticateResponseType {
         authenticatorData: response['authenticatorData'] as String? ?? '',
         signature: response['signature'] as String? ?? '',
         userHandle: (response['userHandle'] as String?) ?? '',
+        clientExtensionResults: json['clientExtensionResults'] as Map<String, Object>?,
       );
     }
 
@@ -48,6 +50,7 @@ class AuthenticateResponseType {
       authenticatorData: json['authenticatorData'] as String? ?? '',
       signature: json['signature'] as String? ?? '',
       userHandle: (json['userHandle'] as String?) ?? '',
+      clientExtensionResults: json['clientExtensionResults'] as Map<String, Object>?,
     );
   }
 
@@ -69,6 +72,9 @@ class AuthenticateResponseType {
   /// The user handle. Can be empty if the user handle is not available.
   final String userHandle;
 
+  /// Optional client extension results.
+  final Map<String?, Object?>? clientExtensionResults;
+
   /// Converts this instance to a JSON string.
   String toJsonString() => jsonEncode(toJson());
 
@@ -76,23 +82,28 @@ class AuthenticateResponseType {
   ///
   /// Output follows the standard WebAuthn format with nested 'response' object.
   Map<String, dynamic> toJson() {
-    final response = <String, dynamic>{
-      'clientDataJSON': clientDataJSON,
-      'authenticatorData': authenticatorData,
-      'signature': signature,
-    };
+    Map<String, dynamic>? sanitizedExtensions;
+
+    final response = <String, dynamic>{'clientDataJSON': clientDataJSON, 'authenticatorData': authenticatorData, 'signature': signature};
 
     // Only include userHandle if it's not empty
     if (userHandle.isNotEmpty) {
       response['userHandle'] = userHandle;
     }
 
-    return {
-      'id': id,
-      'rawId': rawId,
-      'type': 'public-key',
-      'response': response,
-      'clientExtensionResults': <String, dynamic>{},
-    };
+    if (clientExtensionResults != null) {
+      sanitizedExtensions = {};
+
+      final prf = clientExtensionResults?['prf'] as Map?;
+      final results = prf?['results'] as Map?;
+
+      if (results != null) {
+        sanitizedExtensions['prf'] = {
+          'results': {'first': ''},
+        };
+      }
+    }
+
+    return {'id': id, 'rawId': rawId, 'type': 'public-key', 'response': response, 'clientExtensionResults': sanitizedExtensions ?? <String, dynamic>{}};
   }
 }
