@@ -13,8 +13,5 @@ PubKeyCredParamType _$PubKeyCredParamTypeFromJson(Map<String, dynamic> json) =>
     );
 
 Map<String, dynamic> _$PubKeyCredParamTypeToJson(
-        PubKeyCredParamType instance) =>
-    <String, dynamic>{
-      'type': instance.type,
-      'alg': instance.alg,
-    };
+  PubKeyCredParamType instance,
+) => <String, dynamic>{'type': instance.type, 'alg': instance.alg};
