@@ -7,6 +7,7 @@ class RegisterResponseType {
     required this.clientDataJSON,
     required this.attestationObject,
     required this.transports,
+    this.clientExtensionResults,
   });
 
   /// Constructs a new instance from a JSON string.
@@ -22,8 +23,7 @@ class RegisterResponseType {
   factory RegisterResponseType.fromJson(Map<String, dynamic> json) {
     final response = json['response'];
     if (response is! Map<String, dynamic>) {
-      throw FormatException(
-          'Expected "response" to be a Map, got ${response.runtimeType}');
+      throw FormatException('Expected "response" to be a Map, got ${response.runtimeType}');
     }
     final transports = response['transports'] as List<dynamic>?;
 
@@ -33,20 +33,23 @@ class RegisterResponseType {
       clientDataJSON: response['clientDataJSON'] as String? ?? '',
       attestationObject: response['attestationObject'] as String? ?? '',
       transports: transports?.map((e) => e as String?).toList() ?? [],
+      clientExtensionResults: json['clientExtensionResults'] as Map<String, Object>?,
     );
   }
-
   final String id;
   final String rawId;
   final String clientDataJSON;
   final String attestationObject;
   final List<String?> transports;
+  final Map<String?, Object?>? clientExtensionResults;
 
   /// Converts this instance to a JSON string.
   String toJsonString() => jsonEncode(toJson());
 
   /// Converts this instance to a JSON map.
   Map<String, dynamic> toJson() {
+    Map<String, dynamic>? sanitizedExtensions;
+
     final response = <String, dynamic>{
       'clientDataJSON': clientDataJSON,
       'attestationObject': attestationObject,
@@ -58,12 +61,27 @@ class RegisterResponseType {
       response['transports'] = nonNullTransports;
     }
 
+    if (clientExtensionResults != null) {
+      sanitizedExtensions = {};
+
+      final prf = clientExtensionResults?['prf'] as Map?;
+      final results = prf?['results'] as Map?;
+
+      if (results != null) {
+        sanitizedExtensions['prf'] = {
+          'results': {
+            'first': '',
+          }
+        };
+      }
+    }
+
     return {
       'id': id,
       'rawId': rawId,
       'type': 'public-key',
       'response': response,
-      'clientExtensionResults': <String, dynamic>{},
+      'clientExtensionResults': sanitizedExtensions ?? {},
     };
   }
 }
